requirements.txt
selenium==4.23.0
gologin==2024.7.13204038
bs4==0.0.2
beautifulsoup4==4.12.3
mysql.connector==2.2.9
